dataPrediction-kafkaModifying/PintelService/requirements5.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kafka-python==2.0.2
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pykalman==0.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
scipy==1.9.3
six==1.16.0
